Refactor spectrum sim shared logic into header
